refactor: 게임서비스 테스트 코드 클래스 분리
자바 컨벤션 120줄 지키기
